Update and rename MF_Custom-Bounce_v1-6.lua to MF_Custom-Bounce_v1-7.lua
Updated to v1.7, added the ability to select a scene so that objects can be moved in non-current scenes, allowing for much better organization in OBS
